requirements.txt
openai==0.27.6
streamlit==1.25.0
PyPDF2==3.0.1
python-dotenv
langchain==0.0.250
tiktoken==0.4.0
pgml==0.8.1
asyncio==3.4.3